requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2026.1.4
cffi==2.0.0
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.3
daphne==4.1.2
Django==6.0.1
django-allauth==65.4.1
django-tailwind==4.4.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.1.0
pycparser==3.0
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python-slugify==8.0.4
PyYAML==6.0.3
reportlab==4.4.0
requests==2.32.5
rich==14.2.0
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.5
tailwind==3.1.5b0
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
# added dependencies for ASGI and authentication
PyJWT==2.7.0
# daphne is optional; uncomment/install when using channels/ASGI
# daphne==4.0.0

